2012-06-28  Jean Bréfort  <jean.brefort@normalesup.org> 	* configure.ac: add -Werror=format-security to CFLAGS. 	* docs/reference/Makefile.am: add missing files. 	* libs/gcp/application.cc: fixed errors by -Werror=format-security. 	* libs/gcp/window.cc: ditto. 	* libs/gcr/grid.cc: ditto. 	* libs/gcugtk/chem3dapplication.cc: ditto. 	* libs/gcugtk/chem3dwindow.cc: ditto. 	* libs/gcugtk/glview.cc: ditto. 	* libs/gcugtk/message.cc: ditto. 	* libs/gcugtk/message.h: ditto. 	* libs/gcugtk/stringdlg.cc: ditto. 	* plugins/paint/residues/residues-dlg.cc: ditto. 	* plugins/paint/selection/selectiontool.cc: ditto. 	* plugins/paint/templates/templatetool.cc: ditto. 	* programs/calc/gchemcalc.cc: ditto. 	* programs/spectra/application.cc: ditto. 	* programs/table/gchemtable-app.cc: ditto.

+++ libs/gcr/grid.cc
@@ -1338,7 +1338,7 @@ void gcr_grid_delete_row (GcrGrid *grid, unsigned row)
 	std::set < int > decreased;
 	std::set < int >::iterator i, end = grid->selected_rows->end ();
 	for (i = grid->selected_rows->begin (); i != end; i++)
-		if ((*i) > row)
+		if ((*i) > static_cast < int > (row))
 			decreased.insert (*i);
 	grid->selected_rows->erase (row);
 	end = decreased.end ();
